Added support to build using Visual Studio Community (2015)
- Rebuilt libSDL so that it avoided any references to the C runtime library
- Adjusted freetype-2.4.12 to include a locally implemented vsprintf to avoid referencing the C runtime library
- Added logic to pthread.h to only define the timespec structure for Visual Studio Versions prior to 2015

+++ libSDL/SDL2-2.0.3/src/file/SDL_rwops.c
@@ -321,7 +321,7 @@ stdio_seek(SDL_RWops * context, Sint64 offset, int whence)
         return _ftelli64(context->hidden.stdio.fp);
     }
 #else
-    if (fseek(context->hidden.stdio.fp, offset, whence) == 0) {
+    if (fseek(context->hidden.stdio.fp, (long)offset, whence) == 0) {
         return ftell(context->hidden.stdio.fp);
     }
 #endif
